refector: BracketConnectionLayer 파일 분리
